feature:добавлен handler для gRPC

+++ internal/infrastructure/models/models.go
@@ -4,6 +4,6 @@ type Product struct {
 	Id            int     `gorm:"primaryKey;autoIncrement"`
 	Name          string  `gorm:"type:varchar(100);not null"`
 	Description   string  `gorm:"type:varchar(255);"`
-	Price         float64 `gorm:"type:decimal(8,2);not null"`
-	StockQuantity int     `gorm:"type:int;not null"`
+	Price         float32 `gorm:"type:decimal(8,2);not null"`
+	StockQuantity int32   `gorm:"type:int;not null"`
 }
